Implemented edit page for contacts and registration logic

+++ src/main/java/ua/com/javarush/gnew/contactm/repository/EmailRepository.java
@@ -3,5 +3,5 @@ package ua.com.javarush.gnew.contactm.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ua.com.javarush.gnew.contactm.entity.Email;
 
-public interface EmailRepository extends JpaRepository<Email, Integer> {
+public interface EmailRepository extends JpaRepository<Email, Long> {
 }
