+++ src/main/java/ua/com/javarush/gnew/contactm/repository/SocialNetworkRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import ua.com.javarush.gnew.contactm.entity.SocialNetwork;
 
 
-public interface SocialNetworkRepository extends JpaRepository<SocialNetwork, Integer> {
+public interface SocialNetworkRepository extends JpaRepository<SocialNetwork, Long> {
 }
